FST - Add audio to seperate thread
