+++ audio-faust/src/main.rs
@@ -12,75 +12,18 @@
 
 
 mod dsp;
-use dsp::simple_sine::*;
-use cpal::traits::{DeviceTrait, HostTrait, StreamTrait};
+mod audio;
+use audio::*;
 
-fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let host = cpal::default_host();
-    let device = host.default_output_device().expect("No output device available");
-    let mut supported_configs_range = device.supported_output_configs().expect("Error while querying configs");
-    let supported_config = supported_configs_range.next().expect("No supported config?!").with_max_sample_rate();
+use std::io;
 
-    if supported_config.sample_format() != cpal::SampleFormat::F32 {
-        eprintln!("This example requires a device that supports f32 sample format.");
-        return Ok(());
-    }
+fn main()  {
+    start_audio_thread();
+
+    println!("\nAudio is playing. Press Enter to quit.");
     
-    let config = supported_config.config();
-    let sample_rate = config.sample_rate.0;
-    let host_channels = config.channels as usize;
-
-    println!("Audio Device Initialized:");
-    println!("- Sample Rate: {} Hz", sample_rate);
-    println!("- Host Channels: {}", host_channels);
-
-    let mut dsp = SimpleSine::new();
-    dsp.init(sample_rate as i32);
-    let dsp_outputs = dsp.get_num_outputs() as usize;
-    println!("- Faust DSP Channels: {}", dsp_outputs);
-
-    let max_buffer_size = match *device.default_output_config().unwrap().buffer_size() {
-        cpal::SupportedBufferSize::Range { max, .. } => max as usize,
-        cpal::SupportedBufferSize::Unknown => 4096,
-    };
-
-    let mut dsp_output_buffers: Vec<Vec<f32>> = (0..dsp_outputs)
-        .map(|_| vec![0.0; max_buffer_size])
-        .collect();
-
-    let err_fn = |err| eprintln!("an error occurred on the stream: {}", err);
-
-    let stream = device.build_output_stream(
-        &config,
-        move |data: &mut [f32], _: &cpal::OutputCallbackInfo| {
-            let num_frames = data.len() / host_channels;
-
-            let mut dsp_output_slices: Vec<&mut [f32]> = dsp_output_buffers
-                .iter_mut()
-                .map(|buf| &mut buf[..num_frames])
-                .collect();
-
-            let dsp_input_slices: Vec<&[f32]> = Vec::new();
-
-            dsp.compute(num_frames, &dsp_input_slices, &mut dsp_output_slices);
-
-            for i in 0..num_frames {
-                for c in 0..host_channels {
-                    let dsp_channel_index = std::cmp::min(c, dsp_outputs - 1);
-                    let sample = dsp_output_slices[dsp_channel_index][i];
-                    data[i * host_channels + c] = sample;
-                }
-            }
-        },
-        err_fn,
-        None,
-    )?;
-
-    stream.play()?;
-
-    println!("\nPlaying audio, press Enter to quit...");
     let mut input = String::new();
-    std::io::stdin().read_line(&mut input)?;
+    io::stdin().read_line(&mut input).unwrap();
 
-    Ok(())
+    println!("Exiting.");
 }
